Move neuroscope sorting in the same file
