+++ spikeinterface/extractors/extractorlist.py
@@ -8,7 +8,8 @@ from .neoextractors import (
     SpikeGLXRecordingExtractor, read_spikeglx,
     OpenEphysLegacyRecordingExtractor, OpenEphysBinaryRecordingExtractor, OpenEphysBinaryEventExtractor, read_openephys,
     IntanRecordingExtractor, read_intan,
-    NeuroScopeRecordingExtractor, read_neuroscope_recording,
+    NeuroScopeRecordingExtractor, read_neuroscope_recording, NeuroScopeSortingExtractor, read_neuroscope_sorting,
+    read_neuroscope,
     PlexonRecordingExtractor, read_plexon,
     NeuralynxRecordingExtractor, read_neuralynx,
     BlackrockRecordingExtractor, read_blackrock,
@@ -36,7 +37,6 @@ from .spykingcircusextractors import SpykingCircusSortingExtractor, read_spyking
 from .herdingspikesextractors import HerdingspikesSortingExtractor, read_herdingspikes
 from .mdaextractors import MdaRecordingExtractor, MdaSortingExtractor, read_mda_recording, read_mda_sorting
 from .phykilosortextractors import PhySortingExtractor, KiloSortSortingExtractor, read_phy, read_kilosort
-from .neuroscopeextractors import NeuroScopeSortingExtractor, read_neuroscope_sorting
 
 # sorting in relation with simulator
 from .shybridextractors import (SHYBRIDRecordingExtractor, SHYBRIDSortingExtractor,
@@ -72,6 +72,7 @@ recording_extractor_full_list = [
     NixRecordingExtractor,
     NwbRecordingExtractor,
     SpikeGadgetsRecordingExtractor,
+    BiocamRecordingExtractor
 ]
 
 sorting_extractor_full_list = [
